Create main.py
I use www.noip.com for my hostname service.  Not advertising it or getting paid to mention it, just saying it's what I use.  They have a client that updates your hostname to your current public IP address, so this script can query it and return the current IP to the discord chat.  This is used for players to Directly Connect to your Valheim server hosted on your machine.

There ARE security issues here that can be mitigated by Discord permissions settings.  Make sure that the people you give those permissions to are those you trust.  Limit the bot to a single text channel that is only accessible by people with a role that allows them to see that channel.  Don't let just anyone be able to invite to that channel or set roles to access it.