Update main.py
Adding a commands list.  If there is a command using the "!" at the start, the bot will check if it's in the list.  If not, it will inform the user that all commands are in lowercase and to use '!help' to provide a list of available commands.

Not sure the best way to force the incoming message to lowercase.  Would like to eventually be able to split the command on a space so if a user writes "!help you stupid bot" then it only reads "!help".  Currently the command would include the 'you stupid bot' part.