Refactored the main.py file.
Changes:
Refactored the original main.py file to make it use the bot commands.  
Used one command to get all the server info to display to the user.  
Added a way to allow the token to be hidden with a .env file.
Added a connect message from the bot to the channel it's connected to.
When the purge command is used, it clears the channel and re-sends the info message.
TODO:
Add exception handling.