Update main.py
Added rudimentary logging to the terminal as long as the bot is active.  This will show with a timestamp who requested the server information and who cleared the screen.

Updated the info block at the top.

Possible TODO: save the logs to a file on disk.  Not sure if it will be necessary.
Possible TODO: make this a discord bot cog file.